Home Page Update
Cleaned up the Home Page spacing as well as added dynamic content for services. This potentially might cause a problem if there is high traffic on our website. Currently there will be 3 minimal calls, 4 maximum calls, and future (with events) will be 5 maximum calls on each load on the home page. I wonder if we can change this...